refactor(gmx): seperate test setup and transactions
